Improved grammar for feature body and removed ambiguities
The grammar rules for features had to be reworked because they created a serious
ambiguity. The previous rules allowed to define events as features without body.
This clashes with the requirement that local features (including events) should
be possible. E.g. if we have the following feature

   f(a:A)
       local
           i,j := 0,1
           lf(x:X)
               do ...

then it is not clear if the "do" starts the implementation of the local feature lf
or if the local feature lf is just a local event and the "do" starts the implementation
of the feature f. Therefore the resolution is to treat the "do" as the start of the
implementation of the local feature lf and in case that lf is an event it has to be written
as

   f(a:A)
       local
           i,j := 0,1
           lf(x:X)
               note event end
       do ...

In this case the "do" starts the implementation of the outer feature f.

	modified:   ocaml/lexer.mll
	modified:   ocaml/main.ml
	modified:   ocaml/parser.mly
	modified:   ocaml/support.ml