Factored out grammar constructs like "a.b.c.d" from types and expressions
We can have fully qualified types like "a.b.c.d.T". This created ambiguities with
expressions. Therefore the path of a type has been parsed as an expression. This
introduced the full syntactic possibilities of expressions into types. In the
current change a dotted identifier list has been factored out from expressions and
paths. This required a rework of expressions. An expressioin is now either a pexpression
(i.e. a dotted identifier list) or a cexpression (which cannot be a pure dotted identifier
list).

	modified:   parser.mly