Adapted grammar for function expressions
We treat a.f(b) and f(a,b) equivalently. But it is possible that f is an expression
returning a function. In the previous grammar it was possible to express f(a,b) with
a function expression yielding (fexp)(a,b), but the grammar did not allow a.(fexp)(b).
The grammar rules have been adapted to allow this as well.

	modified:   ocaml/lexer.mll
	modified:   ocaml/makefile
	modified:   ocaml/parser.mly
	modified:   ocaml/test.e
	modified:   ocaml/validate.ml