Prover does backward reasoning with the global assertion table
Suppose that the prover wants to prove a goal 'goal' and in the global
assertion table we have the proved assertion

    all(a:BOOLEAN) not not a => a

then the prover loads 'not not goal => goal' into the context. Since
the proved assertion 'not not a => a' is not simplifying in backward
direction a detector is implemented which checks that such a non
simplifying formula cannot by applied repeatedly. Without the
detector the prover would load

    not not goal => goal
    not not not not goal => not not goal
    not not not not not not goal => not not not not goal
    ...

by its backward reasoning engine into the context.

	modified:   library/kernel/boolean.e
	modified:   ocaml/assertion_table.ml
	modified:   ocaml/assertion_table.mli
	modified:   ocaml/container.ml
	modified:   ocaml/prover.ml
	modified:   ocaml/term.ml