Reworked backward reasoning with global assertion table
Backward reasoning with non simplifying rules seem to cause enormeous
combinatorial complexity even if a loop detector is implemented.

Therefore backward reasoning is now restricted to work only if all
subgoals are not more complex (in terms of node count) than the goal
to be proved (provisions are still pending to allow for distributivity).

With this restriction the proof engine works quite fast and good.

The restriction has the disadvantage that the user has to indicate
if he wants an indirect proof (i.e. if he wants to prove 'a' by
assuming 'not a' and deriving a contradiction).

	modified:   library/kernel/boolean.e
	modified:   ocaml/assertion_table.ml
	modified:   ocaml/assertion_table.mli
	modified:   ocaml/container.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli