Forward reasoning with global assertion table implemented
If the assertion table has a proved assertion of the form a=>b and
the context of the prover contains an assertion u which can be unified
with a and the unified variables are sufficient to transform b to an
expression v valid in the environment of u (and v is not more complicated
than u) then u=>v can be added to the context. The local forward engine
of the prover proves then v automatically.

	modified:   library/kernel/boolean.e
	modified:   ocaml/assertion_table.ml
	modified:   ocaml/assertion_table.mli
	modified:   ocaml/prover.ml