Added some code for term elimination (used for case splits)
The module Assertion_table can recognize elimination rules. It stores them and
returns the with the function 'consequences'. In the prover context we can
store elimination rules and a set of eliminated terms to avoid eliminating the
same term over and over.

However the prover does not yet insert elimination terms. Some complexity
problems have to be thought through before.

E.g. if we have the term '(a or b) or c' in the context we get because of the
commutativity of 'or' the term 'c or (a or b)' into the context as well. Both
are elimination terms i.e. for any target we can introduce the two elimination
rules into the context

    (a or b) or c => (a or b => t) => (c => t) => c
    c or (a or b) => (c => t) => (a or b => t) => c

If one of the succeeds then both succeed and if one of the fails both
fail. Therefore in the case of failure the prover would try both options.

With associativity the problem gets worse. Putting '(a or b) or c' into the
context we get the following closure

    (a or b) or c
    c or (a or b)
    a or (b or c)
    (b or c) or a

With equational resoning we could also use subterm commutativity and get
therefore 8 elimination terms into the context. This means in case of failure
we explore 8 routes which are practically equivalent. This combinatorial
explosion has to be avoided.

	modified:   ocaml/assertion_table.ml
	modified:   ocaml/assertion_table.mli
	modified:   ocaml/prover.ml
	modified:   ocaml/term.ml