Restriction of backward reasoning
Consider the following situation

    ..., a=>b=>c                 |- b=>c

In the current implementation two alternatives can be explored

    a) ..., a=>b=>c              |- a

    b) ..., a=>b=>c, b           |- c
       ..., a=>b=>c, b           |- a, b

Both alternatives are equivalent. In the first alternative we have to prove 'a'
under the assumptions '...,a=>b=>c' and in the second we have to prove 'a' and
'b' under the assumptions '...,a=>b=>c,b'. Either both succeed or both fail.

Therefore I have cut off the first alternative from the algorithm. In case of
success the performance is the same. In case of failure the performance is
increased significantly (if there are some of these possibilities).

	modified:   ocaml/prover.ml