Removed redundancies in the backward set
Assume then the prover has the following terms in the context

   a => b => c
   b => a => c

Then the backward set of 'c' consists of {[a,b], [b,a]}. Both alternatives are
equivalent. Either they both succeed or fail to prove 'c'. Therefore only one
of them is added to the backward set of 'c'. This has been done to reduce the
paths to explore if the prover has to prove 'c'.

	modified:   ocaml/prover.ml