Optimized entering
Whenever the goal is an implication chain of the form

    a => b => ... => c

the prover just checks if it can prove the assertion directly (without
backward reasoning) and enters recursively until the goal is 'c' (or one of
the previous direct proofs succeeded). Then it tries to prove 'c' by backward
reasoning. This improved algorithm cuts off some backward reasonings and
therefore improves the performance.

The theoretical basis for this improvement.

    (C |- a => b)  = (C,a |- b)

i.e. if we want to prove 'a=>b' in the context 'C' then we can prove 'b' in
the context 'C' augmented by 'a' as well (and vice versa).

	modified:   ocaml/prover.ml