Fix: Original backward loop breaker fixed
Both loop breaker (the original base on the tried map and the database
approach with the attempt map) are still in the code. Currently the original
one is actived. It has the advantage that it is simpler.

The attempt_map not only detects but also detects if the same goal has already
failed or succeeded earlier. Therefore it generates less goals.

The question remains which of the loop breaker is more efficient from the
globlal perspective. It is a trade off between the number of goals generated
and the effort to maintain the attempt map.

	modified:   ocaml/prover.ml