Reindexing of arguments
Function application is described by the term 'Application(f,args)' where args
is an array and the argument i is args.(i). Since we are using de Bruijn
indices this does not map well because the de Bruijn index of argument i is
n-1-i.

Therefore the order of the argument array has been reversed so that args.(i)
refers to the de Bruijn index i i.e. to the argument n-1-i.

	modified:   ocaml/class_table.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/term.ml