Improved the unifier
If the unifier tries to unify a term which has either a bound variable or a
free variable at a certain position it has access to a map which returns given
the variable directly all assertions (by their indices) which match. No searching
is neccessary.

This improvement speeds up the unifier but slows down to retrieve an assertion
given its index. However the latter operation is practically not needed.

	modified:   ocaml/term_table.ml