Added only the match part of an implication to the backward rules
Assume we have a general rule of the form 'a=>b=>...=>z'. The match
part is the shortest tail which contains all argument variables of the
rule. Only the match part is entered into the set of backward rules of
a general context because the application of a backward rule should not
leave any argument variables open.

A bug in the function 'Term.implication_chain' has been discovered and
fixed

	modified:   ocaml/general_context.ml
	modified:   ocaml/term.ml