Fix: Cumulated argument types missing in 'Local_context'
Local contexts have cumulated argument names but not the corresponding types.
Therefore a local argument is always found in the inner contet, but the type
is taken from the signature which does not contain the type information of
all arguments.

This bug shows up only if we have nested (i.e. more than one) local contexts.

Furthermore the signatures of found global functions are adapted into the local
context. I.e. the index is shifted up appropriately and in the signature there
has to be made space for all formal generics of the local context i.e. between
the formal generics of the global function and the classnames the corresponding
space has to be inserted.

	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/typer.ml