Fix: type variables not updated correctly after the typing of a term
The typing of a term might introduce functions with formal generics.
This implies that type variables which are bound to a constraint are
introduced into the type variables of the type checker. However at
the end of type checking these type variables must be removed,
because they do not belong to the type variables of the context.

Two things can happen if generic functions are used:

a) The formal generics are instantiated with some specific type. This
means that it has to be checked if the specific type satisfies the
constraint.

b) The formal generics can substitute some type variables of the context.
This implies that the free type variables of the context get some
constraints.

Both cases are not yet implemented.

	modified:   ocaml/feature_table.ml
	modified:   ocaml/local_context.ml