Made 'term_table.ml' suited for global and local contexts
Each termtable has now an attribute for the number of variables
in the specific environment. A global environment has zero variables.

For each table node we have the 'cvars' attribute which is a stack of
pairs. Each pair 'n,map' specifies the number of variables in the
corresponding environment and a map which maps variable id's to sets
of assertion id's. The toplevel element of the stack has the greatest
number of variables of the environments in the stack.

	modified:   ocaml/general_context.ml
	modified:   ocaml/term_table.ml
	modified:   ocaml/term_table.mli