Variable names introduced into lambda abstractions
In order to avoid anonymous variable names in lambda abstractions the
variable names have been introduced in lambda abstractions.

Note: The variable types can be recovered from the pure term, but the
variable names not. Therefore it makes sense to include them into the
terms. The names are used only for printouts in order to give the user
the terms how they appear in the source code. The algorithms do not
use the names.

	modified:   ocaml/class_table.ml
	modified:   ocaml/feature_table.ml
	modified:   ocaml/proof.ml
	modified:   ocaml/term.ml
	modified:   ocaml/term_table0.ml