Use backward rules only once
In a context we might have various backward rules for a target

    a => b => t
    c => t

This gives raise to alternatives in the proof search. In the example we have
two alternatives to prove 't'. Either proof 'a' and 'b' or proof 'c'.

During each alternative we might encounter the same target again. If we allow
all backward rules we can use the same rule over and over again.

Therefore if we enter an alternative we mark the corresponding backward rule
as used and don't allow its usage in further down.

	modified:   library/kernel/boolean.e
	modified:   ocaml/local_ass_context.ml
	modified:   ocaml/local_ass_context.mli
	modified:   ocaml/prover.ml
	modified:   boolean1.e