Performance improvement
In the term table we store for all bounded and free variables of a node
already a map to empty substitutions for each assertion id instead of
just the set of assertion ids. Therefore during unification the map is
already available instead of being generated on the fly. This nearly
doubles the performance of the prover.

	modified:   ocaml/term_table0.ml