Completed 'unify_with'
The function 'unify_with' requires for each node to know the subterms
which belong to that node. This is necessary because the node might
match with a variable of the term for which substitutions have to be
found. In this case all the subterms which can substitute the variable
have to be returned.

	modified:   ocaml/term_table0.ml