Infinite loop detection for forward reasoning
Let's assume we have the general rules

    dneg:  all(x)     x => not not x
    and_1: all(x,y)   x and y => x
    and_2: all(x,y)   x and y => y
    curry: all(x,y,z) (x => y => z) => y => x => z

> Remark: We treat `dneg` as a forward rule in order to demonstrate the loop
          breaker. But the prover does not treat `dneg` as a forward rule. But
          the same things can happen in more complicated settings

1. Loop detection with complicating forward rules (`dneg`)

Now we are in a context where we have the proved assertion

    a_1: a and b

Using `dneg` as a forward rule we get

    a_2: a and b => not not (a and b)   {dneg}

We store for each proved assertion the set of general rules which have been
used to derive them. The set of general rules here is `{dneg}`.

Now use modus ponens to derive

    a_3: not not (a and b) {dneg}

Using again `dneg` we can instantiate the forward rule

    a_4: not not (a and b) => not not not not (a and b) {dneg}

We could apply `a_3` and `a_4` with the modus ponens rule, but the prover does
not do that because in both `dneg` has been used as a general rule and `a_4`
is a complicating forward rule.

2. Allowing simplifying forward rules

Let's assume we are in a context where we have the proved assertion

    a_1: a and b and c              -- Note: `and` is left associative

We get the specific forward rules

    a_2: a and b and c => a and b    {and_1}
    a_3: a and b and c => c {and_2}

Applying modus ponens results in

    a_4: a and b {}
    a_5: c {}
       -- Note: `a_2` and `a_3` are simplifying rules. Therefore `a_4` and
       --       `a_5` do not receive the used general rules of `a_2` and
       --       `a_3`.

Because of `a_4` we get the specific forward rules by instantiating `and_1`
and `and_2`.

    a_6: a and b => a  {and_1}
    a_7: a and b => b  {and_2}

Using modus ponens we get

    a_8: a {}
    a_9: b {}

I.e. via forward reasoning the prover can split any assertion of the form
`a and b and c ...` into its elementary parts.

3. The danger of `curry`

Assume we are in a context with the proved assertion

    a_1: a => b => a and b

With that we instantiate the general rule `curry` because `a_1` matches the
premise of `curry`.

    a_2: (a => b => a and b) => b => a => a and b {curry}

and using modus ponens we get

    a_3: b => a => a and b {}   -- `a_2` is not complicating

Note that `a_2` satisfies the pattern of `curry`. This can be seen easier if
put some parentheses.

    a_2: (a => b => a and b) => b => (a => a and b) {curry}

If it were allowed to instatiate `curry` with `a_2` we would get

    a_4: ((a => b => a and b) => b => (a => a and b))
         => b
         => ((a => b => a and b)
             => a => a and b)

I have put some parentheses to see that `a_4` again matches the premise of
`curry`. So we could repeat this process ad infinitum. But the prover does not
insert `a_4` using `curry` because the general rule `curry` has already been
used in `a_2`.

	modified:   ocaml/proof_context.ml