Improvements of the prover
1. Tracing functions improved (now with trace level)

2. Reduced redundancy in backward reasoning

Alternatives can be explored in an nested manner. I.e. during the exploration
of an alternative `i` for a goal `g` (using some backward rule with `g` as a
target and proving the premises) we can encounter the same goal with the same
alternatives except `i`.

If we explore at that point all alternatives different from `i` and fail we
will do the same at the original point of the alternatives.

In order to avoid this I have implemented a checker to see if down in the
stack the prover is already exploring alternatives for the same goal. In this case
it does not explore the alteratives at the top of the stack. I.e. only a breadth
search and no depth search is implemented.

3. Used schematic rules in proved terms

During backward reasoning backward rules can be applied which are not
simplifying. These backward rules are stored in the newly generated goals
(i.e. the subgoals). Therefore the subgoals can be more complicated than the
goal. If the goal is finally proved these used schematic backward rules are
stored with the proved term.

	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/eiffelc.ml
	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli
	modified:   ocaml/prover.ml
	modified:   ocaml/support.ml