Added slots to 'proof_context.ml'
Added some alternative code to make some of the tables based on
TermMap's. The alternative code has been tried on 'find', but no
difference of performance resulted.

The alternative code has been left in, because it might be more
effective on some of the other tables used for backward and forward
reasoning.

	modified:   ocaml/proof_context.ml