Merging formal generics and arguments
Instead of using a name array and a type array for formal arguments
and formal generics we want to use an array of pairs (name,type).

This change includes functions in class_table.ml which work with
array of pairs.

Furthermore the module 'Result_type' has been refactored from 'Sign'
to include the optional result type and information whether it is
a function of a procedure.

	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml
	modified:   ocaml/context.mli
	modified:   ocaml/signature.ml