Implemented typing of untyped predicate or function variables
Typical example

    * (pp:G??): ghost G?
        ensure
            Result = {x: all(p) pp(p) => p(x)}
        end

Suppose the typer is trying to analyze `p(x)`

The required signature is callable i.e. it has `n` arguments and a return type
as requirement.

    (A,B,...):RT        -- required signature

Now assume that the actual variable at this position is untyped i.e. its type
is a type variable without any constraints.

    TV                  -- actual signature

This type variable can be substituted by a function or a predicate with the
correct number of arguments.

But at this point it might not yet be evident if we substitute the variable by
`(A,B,...)->BOOLEAN` or `(A,B,...)?`. We need a type
(e.g. `@DUMMY[(A,B,...),RT]` which can be downgraded to a predicate or a
function if required. I.e. we add the substitution

    TV ~> @DUMMY[(A,B,...),RT]

In order for this to work some other changes were necessary:

- Introduction of a class `@DUMMY` which serves as a placeholder for
  predicates and functions

- Dependency tracking of type variables: Each type variable gets a list of
  other type variables on which it depends (i.e. which occur in its
  substitution). No circularity is allowed. I.e. If we try to add a
  substitution for some type variable `i` we check that none of the used type
  variables in the substitution depends directly or indirectly on `i`.

- The unifier in the typer has to make the appropriate substitutions:

  - If it encounters a fresh type variable in a callable position it creates a
    dummy type with arguments and result type from the required signature.

  - If it encounters a type variable with a substitution in a callable
    position it downgrades the type in the substitution to a callable
    signature and unifies the callable signature with the required signature.

  - If it encounters a type variable with a callable substitution in a
    position with a required constant signature it verifies that the required
    signature is either a predicate or a function, it unifies the arguments
    and possibly the return type and updates the substitution as a predicate
    or a function type.

- Furthermore the typer has to check the typevariables to not contain dummy
  types at certain positions because this would lead to an incomplete inferred
  type.

	modified:   class_table.ml
	modified:   class_table.mli
	modified:   context.ml
	modified:   context.mli
	modified:   signature.ml
	modified:   typer.ml