Fix: Context.add_consequences_implication
If an implication is added to the context it has to be checked if the
there is either an exact match of the premise or there is a schematic
assertion which can be unified with the premise.

The exact matches have been searched in 'prvd2' which works well only if
the premise is not schematic. If the premise is schematic (i.e. all
quantified) the only way to find an exact match is in 'prvd'.

	modified:   context.ml
	modified:   proof_context.ml