Implemented numbers as feature names
Fix: In the typer new untyped type variables are always added at the
bottom and new global type variables (from formal generics of global
features) are always added at the bottom. During unification a newer
type variable can be substituted by an older or a conceptless by a
variable with concept. The other way round is illegal. This order is
enforced now.

	modified:   context.ml
	modified:   parser.mly
	modified:   signature.ml
	modified:   support.ml
	modified:   typer.ml