Splitted up the formal generics array
Instead of using '(int*type_term) array' to represent the formal generics
we use two arrays, one for the names and one for the concepts.

Reason: There are functions which use only the concepts or only the names
and functions which have access only to one of the arrays but need to call
some other function which used to receive both without necessity.

	modified:   class_table.ml
	modified:   class_table.mli
	modified:   context.ml
	modified:   feature_table.ml
	modified:   feature_table.mli