Ownership calculation simplified
Each signature is owned by a class. Since all class indices and all
module indices are now nonnegative integers the ownership calculation
could be simplified.

In order to calculate the ownership of a signature one has to compute
the set of all involved classes. The class with the highest module
rank is the owner of the signature.

Not yet implemented: Modules with multiple classes!!

	modified:   class_table.ml
	modified:   class_table.mli
	modified:   feature_table.ml
	modified:   signature.ml