Concepts of type variables (formal generics) no longer absolute
'TVars.t' is a context for types and contains the number of free type
variables, for each globally imported (from the feature table) type variable a
concept and for each formal generic of the context a concept.

Up to now the concepts have been absolute. Therefore they could not refer to
other type variables in the type context.

Now all concepts (of imported or local type variables) are relative to the
type context. With this adaption concepts of type variables can refer to other
type variables.

With this relativeness we can define formal generics which depend on other
formal generics.

Example:

    G: ANY

    deferred class ABSTRACT_ARRAY[G] ... end

    AA[G]: ABSTRACT_ARRAY[G]    -- not yet accepted by the parser

    size(a:AA[G]): NATURAL      deferred end

    [] (a:AA[G], i:NATURAL): G
        require
            i < a.size
        deferred end

    put(a:AA[G], i:NATURAL, v:G)
        require
             i < a.size
        deferred ensure
             a[i] = v
        end

	modified:   class_table.ml
	modified:   context.ml
	modified:   context.mli
	modified:   feature_table.ml
	modified:   signature.ml
	modified:   term_builder.ml