Added the function 'TVars.is_equal'
The function 'is_equal tp1 tvs1 tp2 tvs2' compares the two types 'tp1'
and 'tp2' which come from different environments 'tvs1' and 'tvs2'. Two
types are considered equal if they have the same classes in the same
positions of the type term or both a type variable in the same position
and their corresponding concepts are equal.

	modified:   tvars.ml