Fix: Discharged term must be transformed to the local environment
The prover might encounter the fact that the current goal has already
been proved in some outer environment. Before discharging this term
with the local assumptions and all-quantifying it, the term has to
be transformed into the local environment.

This bug has not been detected up to now because in the most cases the
discharged term is from the local environment.

	modified:   proof_table.ml