Seeds and variants implemented in feature table
Initially each feature is its own seed and has no variants.

If a feature is inherited in a descendant, then the inherited feature
inherits all seeds from the feature and is no longer its own seed
and all seeds of the feature have the feature in the descendant as
a variant.

	modified:   feature_table.ml