'Term_builder.update_term' implemented and used for inheritance
The term builder finds first the most general term using e.g.
deferred features. The function 'update_term' checks if there is a
more specific feature in the descendant class and uses this if
possible instead of the most general feature.

The most specific features do not appear in the feature map. Therefore
only the most general features (i.e. the seeds) are found first. The
update procedure finds the correct variant.

	modified:   ocaml/feature_table.ml
	modified:   ocaml/feature_table.mli
	modified:   ocaml/signature.ml
	modified:   ocaml/term_builder.ml
	modified:   ocaml/term_builder.mli
	modified:   ocaml/term_table.ml
	modified:   ocaml/term_table.mli
	modified:   ocaml/typer.ml