Equality reasoning with leibniz axiom and predicates possible
The functionality is best explained with an example. The module 'predicate'
contains the leibniz axiom

    all(a,b:G)            -- G:ANY
        require  a = b
        note     built_in
        ensure   all(p:G?) p(a) => p(b) end

In simple words: Whatever can be said about 'a', when 'a=b' is valid applies
to 'b' as well.

The compiler ensures that any definition of equality in a descendant of ANY
will satisfy this axiom.

The module 'any' just requires reflexivity of '='. But with the leibniz axiom
the following proof is possible:

    all(a,b:G)
        require   a = b
        check     {x: x=a}(b)   -- c0
        ensure    b = a         end

The prover makes the following steps. On seeing 'a=b' it puts deriving from
the leibniz axiom the assertion 'all(p) p(a) => p(b)' into the context. Now it
matches 'p(b)' with '{x:x=a}(b)' and puts '{x:x=a}(a) => {x:x=a}(b)' into the
context. But before putting it into the context it makes a beta
reduction. Therefore it puts 'a=b => b=a' as a backward rule into the context
which proves 'c0' and the final goal trivially.

In order to achieve this functionality the following rules have to be adhered:

- Only beta reduced assertions are put into the context.

- On finding backward rules the prover checks first the possible backward rules
  before reduction, adds them to the context and then beta reduces the goal.

- The proof_table and the proof context have to make sure that addition of
  specializations can result in redexes which have to be reduced before
  insertion.

	modified:   context.ml
	modified:   proof_context.ml
	modified:   proof_table.ml
	modified:   prover.ml