Fix: sequence of type variable substitutions
The type introduces new type variables into the type context in two ways:

- If it expects a function it puts a new conceptless type variable at the
  bottom of the type variables for each argument of the function

- If global functions are encountered which are generic it puts new type
  variables with concepts at the top of the type variables

Usually newer type variables get substituted by older. This is always the case
for conceptless type variables because they get a substitution before new
conceptless type variables are introduced.

It is not always the case for type variables with concepts. The default order
might be reversed if the concept of the older is satisfied by the concept of
the newer and not vice versa. Then the older will be substituted by the newer.

	modified:   ocaml/signature.ml
	modified:   ocaml/term_builder.ml