Inhibit inheritance to public classes in implementation files
Inheritance to public classes (i.e. classes which have been defined in
some of the used modules) is not allowed in implementation files, because
the inheritance puts classes as ancestors to publicly visible classes which
users of the implemented module might not be able to see.

Note: This strict rule might be made more specific in the future.

Furthermore: Removal of unneeded functions.

	modified:   ocaml/class_table.ml
	modified:   ocaml/class_table.mli
	modified:   ocaml/context.ml