Fix: unifier in term table
The term table has a set of schematic terms i.e. terms with variables which
can be substituted.

Unification is done by checking a term to be unified from top to bottom
according to its structure. Whenever there is a variable of a schematic term
in the table at the corresponding position a substitution is generated for the
variable. However this is incorrect if the substitution contains bound
variables of the term to be checked. No term with bound variables can
substitute a variable!

	modified:   ocaml/term_table.ml