Allowing backward rules which do not simplify and not introduce
Now all backward rules are allowed as long as the target is the shortest
tail which contains all argument variables and is simplifying or not a single
variable (no catchall rules i.e. still no automatic indirect proofs possible).

The backward rules are still classified as simplifying (i.e. the node count
of any premise is not greater than the node count of the target) and non
simplifying.

The non simplifying rules cannot be used repeately to avoid infinite loops
in the verifier.

	modified:   ocaml/proof_context.ml