Fix: find_equivalent -> find_stronger
An assertion shall not be entered into the proof tables if there is
an equivalent or stronger assertion already in the context.

An assertion is equivalent or stronger if it has a less or equal number
of variables and the stronger assertion can be transformed into the
weaker assertion by using an injective substitution of the variables.

An assertion is weaker in this sense only if it doesn't use some of its
variables. Such an assertion is usually useless for the prover.

	modified:   ocaml/proof_context.ml
	modified:   ocaml/proof_context.mli